chore: rename files Signed-off-by: Sonia Sandler <ssandler@redhat.com>
