+++ packages/renderer/src/lib/statusbar/ProviderTooltipStatusStyle.svelte
